added Tensor constructor

+++ src/tensor.cc
@@ -4,9 +4,11 @@ namespace jb {
 
 namespace tensor {
 
+// UTILITY FUNCTIONS
+
 vector<int> ShapeToStrides(const vector<int> &shape) {
-  int ndim = shape.size();
-  vector<int> strides(ndim);
+  int ndim = (int)shape.size();
+  vector<int> strides((int)ndim);
   strides[ndim - 1] = 1;
   for (int i = ndim - 2; i >= 0; i--) {
     strides[i] = shape[i + 1] * strides[i + 1];
@@ -14,6 +16,8 @@ vector<int> ShapeToStrides(const vector<int> &shape) {
   return strides;
 }
 
-}
+// TENSOR CLASS
+
+}  // namespace tensor
 
 }  // namespace jb
